Code moved to SlaveComputer class
