+++ core/src/main/java/hudson/model/Computer.java
@@ -29,7 +29,6 @@ import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import hudson.EnvVars;
 import hudson.Extension;
 import hudson.Launcher.ProcStarter;
-import hudson.Main;
 import hudson.slaves.Cloud;
 import jenkins.security.stapler.StaplerDispatchable;
 import hudson.Util;
@@ -1175,32 +1174,6 @@ public /*transient*/ abstract class Computer extends Actionable implements Acces
     @Deprecated
     public Map<String,String> getEnvVars() throws IOException, InterruptedException {
         return getEnvironment();
-    }
-
-    /**
-     * Expose real full env vars map from agent for UI presentation
-     */
-    public Map<String,String> getEnvVarsFull() throws IOException, InterruptedException {
-        if(getChannel() == null) {
-            Map<String, String> env = new TreeMap<> ();
-            env.put("N/A","N/A");
-            return env;
-        } else {
-            return getChannel().call(new ListFullEnvironment());
-        }
-    }
-
-    private static class ListFullEnvironment extends MasterToSlaveCallable<Map<String,String>,IOException> {
-        public Map<String,String> call() throws IOException {
-            Map<String, String> env = new TreeMap(System.getenv());
-            if(Main.isUnitTest || Main.isDevelopmentMode) {
-                // if unit test is launched with maven debug switch,
-                // we need to prevent forked Maven processes from seeing it, or else
-                // they'll hang
-                env.remove("MAVEN_OPTS");
-            }
-            return env;
-        }
     }
 
     /**
